requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-chroma
sentence-transformers
python-dotenv